+++ envi/qt/memwrite.py
@@ -263,7 +263,7 @@ class MemWriteWindow(QtGui.QWidget):
 
 class MockEmu(object):
     def parseExpression(self, expr):
-        return long(eval(expr, {}, {}))
+        return int(eval(expr, {}, {}))
 
     def readMemory(self, va, size):
         return '\x90' * size
