+++ vdb/extensions/winkern.py
@@ -147,7 +147,7 @@ def bugcheck(db, line):
 
                 sym = trace.getSymByAddr(eip, exact=False)
                 if sym != None:
-                    framestr += ' (%s + %d)' % (sym,eip-long(sym))
+                    framestr += ' (%s + %d)' % (sym,eip-int(sym))
 
                 db.vprint(framestr)
 
@@ -162,7 +162,7 @@ def bugcheck(db, line):
             sym = trace.getSymByAddr(exaddr, exact=False)
             db.vprint('')
             db.vprint('Exception Trap: 0x%.8x' % extrap)
-            db.vprint('Exception Addr: 0x%.8x (%s + %d)' % (exaddr, sym, exaddr-long(sym)))
+            db.vprint('Exception Addr: 0x%.8x (%s + %d)' % (exaddr, sym, exaddr-int(sym)))
             db.vprint('Exception Code: 0x%.8x' % excode)
             db.vprint('Exception Inst: %r' % trace.parseOpcode(exaddr))
